separate 'snet account' logic
